backend/requirements.txt

Flask
Flask-Cors
python-dotenv
Flask-SQLAlchemy
flask-migrate
PyJWT
Werkzeug
flask-mail
APScheduler

PyMuPDF
docx2txt
python-docx

openai
langchain
langchain-openai
langchain-community
langgraph

gtts
pydub
SpeechRecognition
pyaudio


openai-whisper
ffmpeg-python
